[docs] Generate linting docs from source
When new rules have been added to WPT's "lint" tool, the corresponding
documentation has not always been updated [1] [2] [3]. Automatically
generating documentation from source code helps avoid this state and the
confusion it can cause contributors.

Introduce a pattern for defining linting rules. Rely on the new
structure during documentation generation to automatically create a
listing of all available linting rules.

The new structure centralizes information about the rules (name,
description, and instructions for correcting infractions). The process
honors the existing convention of Markdown within descriptions.

Although the Sphinx documentation generator includes a built-in
extension for generating documentation from Python source code, the
output of that extension is designed to document Python primitives such
as functions and classes. Such a format is inappropriate for this case
because the users of the linting tool do not interact with the internals
in this way. Define a custom docutils directive to tailor the
documentation to the needs of its audience.

[1] https://github.com/web-platform-tests/wpt/issues/5299
[2] https://github.com/web-platform-tests/wpt/issues/10501
[3] https://github.com/web-platform-tests/wpt/issues/11479